Move discover module under rele.management
- Django believes this module is another command
